refactor: renames store/context.tsx to store/provider.tsx and update import paths
